pyproject.toml
[tool.poetry]
name = "policyparser"
version = "0.1.0"
description = ""
authors = ["PermuteHQ <contact@permutehq.com>"]

[tool.poetry.dependencies]
python = "^3.10"
PyPDF2 = "^2.11.2"
defusedxml = "^0.7.1"
lxml = "^4.9.1"
openpyxl = "^3.0.10"
python-Levenshtein = "^0.20.8"
thefuzz = "^0.19.0"
"pdfminer.six" = "^20221105"

[tool.poetry.dev-dependencies]
black = {version = "^22.3.0", allow-prereleases = true}
flake8 = "^5.0.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
